admin-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.0
pydantic-settings==2.2.1
supabase==2.4.0
python-multipart==0.0.9
razorpay==1.5.0
